# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SmartSchool Asset API",
        "description": "Dokumentasi API untuk sistem manajemen aset SmartSchool. Mencakup peminjaman barang, pengambilan konsumsi, barang masuk, dan laporan.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local Development Server"
        }
    ],
    "paths": {
...
        "/cabinets": {
            "get": {
                "tags": [
                    "Lemari"
                ],
                "summary": "Daftar semua lemari (admin)",
                "description": "Menampilkan daftar semua lemari beserta ruangan tempatnya berada. Hanya untuk admin.",
                "operationId": "6db282f88b4f1cb20244726f76f318f9",
                "responses": {
                    "200": {
                        "description": "Halaman daftar lemari"
                    },
                    "403": {
                        "description": "Akses ditolak"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Lemari"
                ],
                "summary": "Simpan lemari baru (admin)",
                "description": "Menyimpan data lemari baru ke sistem. Kode lemari harus unik dan ruangan wajib dipilih.",
                "operationId": "45daa56e96aa32bb3382b533ff62bd66",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "kode_lemari",
                                    "nama_lemari",
                                    "id_ruangan"
                                ],
                                "properties": {
                                    "kode_lemari": {
                                        "description": "Kode unik lemari",
                                        "type": "string",
                                        "example": "L-01"
                                    },
                                    "nama_lemari": {
                                        "type": "string",
                                        "example": "Lemari Elektronik"
                                    },
                                    "id_ruangan": {
                                        "description": "ID ruangan tempat lemari berada",
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "302": {
                        "description": "Redirect ke daftar lemari jika berhasil"
                    },
                    "422": {
                        "description": "Validasi gagal (kode sudah dipakai atau ruangan tidak ada)"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/cabinets/create": {
            "get": {
                "tags": [
                    "Lemari"
                ],
                "summary": "Form tambah lemari (admin)",
                "description": "Menampilkan form untuk menambahkan lemari baru. Hanya untuk admin.",
                "operationId": "1f3d83b828c9ab14c354ea692583ad42",
                "responses": {
                    "200": {
                        "description": "Halaman form tambah lemari"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/cabinets/{id}/edit": {
            "get": {
                "tags": [
                    "Lemari"
                ],
                "summary": "Form edit lemari (admin)",
                "description": "Menampilkan form untuk mengedit data lemari. Hanya untuk admin.",
                "operationId": "9eea6c957005a727b857baf0470c2055",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Halaman form edit lemari"
                    },
                    "404": {
                        "description": "Lemari tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/cabinets/{id}": {
            "put": {
                "tags": [
                    "Lemari"
                ],
                "summary": "Update data lemari (admin)",
                "description": "Mengupdate data lemari. Kode lemari harus unik kecuali milik lemari itu sendiri.",
                "operationId": "c393371d19fe9f416c96b46fcd2f3d84",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "kode_lemari",
                                    "nama_lemari",
                                    "id_ruangan"
                                ],
                                "properties": {
                                    "kode_lemari": {
                                        "type": "string",
                                        "example": "L-01"
                                    },
                                    "nama_lemari": {
                                        "type": "string",
                                        "example": "Lemari ATK"
                                    },
                                    "id_ruangan": {
                                        "type": "integer",
                                        "example": 2
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "302": {
                        "description": "Redirect ke daftar lemari jika berhasil"
                    },
                    "422": {
                        "description": "Validasi gagal"
                    },
                    "404": {
                        "description": "Lemari tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Lemari"
                ],
                "summary": "Hapus lemari (admin)",
                "description": "Menghapus data lemari dari sistem. Hanya untuk admin.",
                "operationId": "e06934492e8df09b32010e4296a9f5ed",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "302": {
                        "description": "Redirect dengan pesan sukses"
                    },
                    "404": {
                        "description": "Lemari tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Masukkan token autentikasi Laravel Sanctum/Session",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}